requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.0
qrcode==7.4.2
pillow==10.1.0
jinja2==3.1.2
python-dotenv==1.0.0
aiosmtplib==3.0.1
